requirements.txt
Django==2.0.1
django-taggit==0.22.2
Pillow==5.0.0
pytz==2017.3
